refactor models, first shot at relationships
for now we are duplicating the name of the trait in the frame instead
of actually creating a proper relationship between the trait and the
frame, mainly because I can't wrap my stupid head around the way that
datamapper structures relationships. Every time I think it should work
it doesn't. In an ideal world, a Frame would have 1 Trait and a Trait
would have many Frames, but this doesn't seem to float with DM, at least
not in the procedural order I expect it to, and so I'm going to have to
figure things out. Note to future self: this blog post seems to be the
best thing there is out there http://bit.ly/ctgvAd